+++ backend/src/main/java/com/celuveat/celuveat/celeb/domain/Celeb.java
@@ -1,4 +1,4 @@
-package com.celuveat.celuveat.celeb;
+package com.celuveat.celuveat.celeb.domain;
 
 import lombok.Builder;
