+++ backend/src/main/java/com/celuveat/celuveat/celeb/infra/persistence/CelebDao.java
@@ -1,7 +1,9 @@
-package com.celuveat.celuveat.video;
+package com.celuveat.celuveat.celeb.infra.persistence;
 
-import static com.celuveat.celuveat.video.VideoExceptionType.NOT_FOUND_VIDEO;
+import static com.celuveat.celuveat.celeb.exception.CelebExceptionType.NOT_FOUND_CELEB;
 
+import com.celuveat.celuveat.celeb.domain.Celeb;
+import com.celuveat.celuveat.celeb.exception.CelebException;
 import com.celuveat.celuveat.common.annotation.Dao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.DataClassRowMapper;
@@ -12,36 +14,36 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 
 @Dao
-public class VideoDao {
+public class CelebDao {
 
-    private static final RowMapper<Video> mapper = new DataClassRowMapper<>(Video.class);
+    private static final RowMapper<Celeb> mapper = new DataClassRowMapper<>(Celeb.class);
 
     private final JdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert simpleJdbcInsert;
 
-    public VideoDao(JdbcTemplate jdbcTemplate) {
+    public CelebDao(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
         this.simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("video")
+                .withTableName("celeb")
                 .usingGeneratedKeyColumns("id");
     }
 
-    public Long save(Video video) {
-        SqlParameterSource source = new BeanPropertySqlParameterSource(video);
+    public Long save(Celeb celeb) {
+        SqlParameterSource source = new BeanPropertySqlParameterSource(celeb);
         return simpleJdbcInsert.executeAndReturnKey(source)
                 .longValue();
     }
 
-    public Video getById(Long id) {
+    public Celeb getById(Long id) {
         String sql = """
-                SELECT * 
-                FROM video
+                SELECT *
+                FROM celeb
                 WHERE id = ?
                 """;
         try {
             return jdbcTemplate.queryForObject(sql, mapper, id);
         } catch (EmptyResultDataAccessException e) {
-            throw new VideoException(NOT_FOUND_VIDEO);
+            throw new CelebException(NOT_FOUND_CELEB);
         }
     }
 }
