--- backend/src/main/java/com/celuveat/celuveat/restaurant/Category.java
@@ -1,7 +1,0 @@
-package com.celuveat.celuveat.restaurant;
-
-public enum Category {
-
-    ETC,
-    ;
-}
