+++ backend/src/main/java/com/celuveat/celuveat/restaurant/domain/Category.java
@@ -1,0 +1,7 @@
+package com.celuveat.celuveat.restaurant.domain;
+
+public enum Category {
+
+    ETC,
+    ;
+}
