+++ backend/src/main/java/com/celuveat/celuveat/restaurant/exception/RestaurantException.java
@@ -1,13 +1,13 @@
-package com.celuveat.celuveat.video;
+package com.celuveat.celuveat.restaurant.exception;
 
 import com.celuveat.celuveat.common.exception.BaseException;
 import com.celuveat.celuveat.common.exception.BaseExceptionType;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public class VideoException extends BaseException {
+public class RestaurantException extends BaseException {
 
-    private final VideoExceptionType exceptionType;
+    private final RestaurantExceptionType exceptionType;
 
     @Override
     public BaseExceptionType exceptionType() {
