+++ backend/src/main/java/com/celuveat/celuveat/restaurant/infra/persistence/RestaurantDao.java
@@ -1,8 +1,10 @@
-package com.celuveat.celuveat.celeb;
+package com.celuveat.celuveat.restaurant.infra.persistence;
 
-import static com.celuveat.celuveat.celeb.CelebExceptionType.NOT_FOUND_CELEB;
+import static com.celuveat.celuveat.restaurant.exception.RestaurantExceptionType.NOT_FOUND_RESTAURANT;
 
 import com.celuveat.celuveat.common.annotation.Dao;
+import com.celuveat.celuveat.restaurant.domain.Restaurant;
+import com.celuveat.celuveat.restaurant.exception.RestaurantException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.DataClassRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -12,36 +14,36 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 
 @Dao
-public class CelebDao {
+public class RestaurantDao {
 
-    private static final RowMapper<Celeb> mapper = new DataClassRowMapper<>(Celeb.class);
+    private static final RowMapper<Restaurant> mapper = new DataClassRowMapper<>(Restaurant.class);
 
     private final JdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert simpleJdbcInsert;
 
-    public CelebDao(JdbcTemplate jdbcTemplate) {
+    public RestaurantDao(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
         this.simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("celeb")
+                .withTableName("restaurant")
                 .usingGeneratedKeyColumns("id");
     }
 
-    public Long save(Celeb celeb) {
-        SqlParameterSource source = new BeanPropertySqlParameterSource(celeb);
+    public Long save(Restaurant restaurant) {
+        SqlParameterSource source = new BeanPropertySqlParameterSource(restaurant);
         return simpleJdbcInsert.executeAndReturnKey(source)
                 .longValue();
     }
 
-    public Celeb getById(Long id) {
+    public Restaurant getById(Long id) {
         String sql = """
                 SELECT *
-                FROM celeb
+                FROM restaurant
                 WHERE id = ?
                 """;
         try {
             return jdbcTemplate.queryForObject(sql, mapper, id);
         } catch (EmptyResultDataAccessException e) {
-            throw new CelebException(NOT_FOUND_CELEB);
+            throw new RestaurantException(NOT_FOUND_RESTAURANT);
         }
     }
 }
