+++ backend/src/main/java/com/celuveat/celuveat/video/exception/VideoException.java
@@ -1,13 +1,13 @@
-package com.celuveat.celuveat.restaurant;
+package com.celuveat.celuveat.video.exception;
 
 import com.celuveat.celuveat.common.exception.BaseException;
 import com.celuveat.celuveat.common.exception.BaseExceptionType;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public class RestaurantException extends BaseException {
+public class VideoException extends BaseException {
 
-    private final RestaurantExceptionType exceptionType;
+    private final VideoExceptionType exceptionType;
 
     @Override
     public BaseExceptionType exceptionType() {
