+++ backend/src/main/java/com/celuveat/celuveat/video/exception/VideoExceptionType.java
@@ -1,18 +1,18 @@
-package com.celuveat.celuveat.restaurant;
+package com.celuveat.celuveat.video.exception;
 
 import com.celuveat.celuveat.common.exception.BaseExceptionType;
 import org.springframework.http.HttpStatus;
 
-public enum RestaurantExceptionType implements BaseExceptionType {
+public enum VideoExceptionType implements BaseExceptionType {
 
-    NOT_FOUND_RESTAURANT(200, HttpStatus.NOT_FOUND, "음식점을 찾을 수 없습니다"),
+    NOT_FOUND_VIDEO(300, HttpStatus.NOT_FOUND, "영상을 찾을 수 없습니다"),
     ;
 
     private final int errorCode;
     private final HttpStatus httpStatus;
     private final String errorMessage;
 
-    RestaurantExceptionType(int errorCode, HttpStatus httpStatus, String errorMessage) {
+    VideoExceptionType(int errorCode, HttpStatus httpStatus, String errorMessage) {
         this.errorCode = errorCode;
         this.httpStatus = httpStatus;
         this.errorMessage = errorMessage;
